wicdec: silence a format warning
from x86_64-w64-mingw32-gcc

examples/wicdec.c: In function ‘ExtractICCP’:
examples/wicdec.c:131:21: warning: format ‘%u’ expects argument of type
‘unsigned int’, but argument 4 has type ‘size_t’ [-Wformat]

(cherry picked from commit ffae9f31e80ed7344987b24b21af831806fa2e69)

+++ examples/wicdec.c
@@ -128,7 +128,7 @@ static HRESULT ExtractICCP(IWICImagingFactory* const factory,
                                                &size));
           if (SUCCEEDED(hr) && size != iccp->size) {
             fprintf(stderr, "Warning! ICC profile size (%u) != expected (%u)\n",
-                    size, iccp->size);
+                    size, (uint32_t)iccp->size);
             iccp->size = size;
           }
           break;
